Fixed bug causing error with API tests.
The error with the API tests was caused if the target filestore location
has not been created when the tests are run. For the temporary uploads,
this was not an issue since the temporary upload directory was being created
within the test class setup. Since these directories will be required by the
libraries, the library initialisation has been modified to create the required
directories when the library is initialised. The previous directory creation
code has been removed from the test classes since this is no longer required.